move logic to adapter
