--- inlang/packages/paraglide/paraglide-next/eslint.config.js
@@ -1,13 +1,0 @@
-import pluginJs from "@eslint/js";
-import tseslint from "typescript-eslint";
-
-export default [
-	pluginJs.configs.recommended,
-	...tseslint.configs.recommended,
-	{
-		rules: {
-			// any makes testing sometimes easier
-			"@typescript-eslint/no-explicit-any": "off",
-		},
-	},
-];
