+++ inlang/packages/paraglide/paraglide-next/example/package.json
@@ -14,6 +14,7 @@
 	},
 	"devDependencies": {
 		"@inlang/paraglide-js": "workspace:*",
+		"@inlang/paraglide-next": "workspace:*",
 		"typescript": "^5",
 		"@types/node": "^20",
 		"@types/react": "^19",
