+++ inlang/packages/paraglide/paraglide-next/example/src/app/ClientSideLocaleSwitch.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { setLocale } from "./paraglide/runtime";
+import { setLocale } from "../paraglide/runtime";
 
 export default function ClientSideLocaleSwitch() {
 	return (
