+++ inlang/packages/paraglide/paraglide-next/example/src/app/page.tsx
@@ -1,4 +1,4 @@
-import * as m from "./paraglide/messages.js";
+import * as m from "../paraglide/messages.js";
 
 export default function Home() {
 	return <p>{m.example_message({ username: "Samuel" })}</p>;
