+++ inlang/packages/paraglide/paraglide-next/example/src/middleware.ts
@@ -1,5 +1,5 @@
 import { NextRequest } from "next/server";
-import * as paraglide from "./app/adapter.js";
+import * as paraglide from "./paraglide/adapter";
 
 export function middleware(request: NextRequest) {
 	return paraglide.middleware(request);
