+++ inlang/packages/paraglide/paraglide-next/package.json
@@ -14,41 +14,33 @@
 		"url": "https://github.com/opral/inlang-paraglide-js"
 	},
 	"files": [
-		"./dist",
-		"./bin",
-		"./default"
+		"./dist"
 	],
 	"exports": {
-		".": {
-			"import": "./dist/index.js",
-			"types": "./dist/index.d.ts",
-			"svelte": "./dist/index.js"
-		}
+		".": "./dist/index.js"
 	},
 	"scripts": {
 		"dev": "tsc --watch",
 		"build": "tsc",
 		"test": "tsc --noEmit && vitest run --coverage ./src/**/* --passWithNoTests",
 		"test:watch": "npm run env-variables && vitest --watch ./src/**/*",
-		"lint": "eslint ./src --fix",
 		"format": "prettier ./src --write",
 		"clean": "rm -rf ./dist ./node_modules"
 	},
 	"dependencies": {},
 	"peerDependencies": {
-		"@inlang/paraglide-js": "^2.0.0"
+		"@inlang/paraglide-js": "^2.0.0",
+		"next": "^15"
 	},
 	"devDependencies": {
 		"@inlang/paraglide-js": "workspace:*",
-		"@eslint/js": "^9.18.0",
 		"@opral/tsconfig": "workspace:*",
 		"@types/node": "^22.10.6",
 		"@vitest/coverage-v8": "2.1.8",
-		"eslint": "^9.18.0",
 		"prettier": "^3.4.2",
-		"typescript-eslint": "^8.20.0",
 		"typescript": "^5.7.3",
-		"vitest": "2.1.8"
+		"vitest": "2.1.8",
+		"next": "^15"
 	},
 	"keywords": [
 		"inlang",
