+++ inlang/packages/paraglide/paraglide-next/tsconfig.json
@@ -1,10 +1,15 @@
 {
 	"extends": "@opral/tsconfig/default",
-	"include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.js"],
+	"include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.js", "src/**/*.jsx"],
 	"compilerOptions": {
+		"paths": {
+			"next/server": ["./node_modules/next/server.d.ts"],
+			"next/headers": ["./node_modules/next/headers.d.ts"]
+		},
 		"outDir": "./dist",
 		"rootDir": "./src",
 		"module": "Node16",
+		"jsx": "preserve",
 		"declaration": true,
 		"sourceMap": false,
 		"checkJs": true,
